lift topn pruning
